+++ lib/routes/amazfitwatchfaces/top.js
@@ -1,47 +1,9 @@
-const url = require('url');
-const got = require('@/utils/got');
-const cheerio = require('cheerio');
+const utils = require('./utils');
 
 module.exports = async (ctx) => {
-    const rootUrl = `https://amazfitwatchfaces.com/`;
-    const currentUrl = `${rootUrl}${ctx.params.model}/top?${ctx.params.lang ? 'lang=' + ctx.params.lang : ''}${ctx.params.type ? '&compatible=' + ctx.params.type : ''}${ctx.params.sortBy ? '&sortby=' + ctx.params.sortBy : ''}${
+    const currentUrl = `${ctx.params.model}/top?${ctx.params.lang ? 'lang=' + ctx.params.lang : ''}${ctx.params.type ? '&compatible=' + ctx.params.type : ''}${ctx.params.sortBy ? '&sortby=' + ctx.params.sortBy : ''}${
         ctx.params.time ? '&topof=' + ctx.params.time : ''
     }`;
-    const response = await got({
-        method: 'get',
-        url: currentUrl,
-    });
-    const $ = cheerio.load(response.data);
-    const list = $('#wf-row div.col-md-3')
-        .slice(0, 10)
-        .map((_, item) => {
-            item = $(item);
-            return {
-                link: url.resolve(rootUrl, $(item).find('a.wf-act').attr('href')),
-            };
-        })
-        .get();
 
-    const items = await Promise.all(
-        list.map(
-            async (item) =>
-                await ctx.cache.tryGet(item.link, async () => {
-                    const detailResponse = await got({ method: 'get', url: item.link });
-                    const content = cheerio.load(detailResponse.data);
-                    const date = content('i.fa-calendar').parent().find('span').text();
-                    const ymd = date.split(' ')[0];
-
-                    item.title = `[${content('code').text()}] ${content('title').text().split(' - ')[0]}`;
-                    item.description = `<img src="${content('#watchface-preview').attr('src')}"><br>${content('div.unicodebidi').text()}`;
-                    item.pubDate = new Date(`${ymd.split('.')[2]}-${ymd.split('.')[1]}-${ymd.split('.')[0]} ${date.split(' ')[1]}`).toUTCString();
-                    return item;
-                })
-        )
-    );
-
-    ctx.state.data = {
-        title: `${$('title').text().split('|')[0]} | Amazfit Watch Faces`,
-        link: currentUrl,
-        item: items,
-    };
+    ctx.state.data = await utils(ctx, currentUrl);
 };
